+++ CHANGELOG.md
@@ -64,6 +64,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   other can be referred to as `xyz/1`. These suffixes are shown in the log when
   necessary to make a change ID unambiguous.
 
+* `jj util gc` now prunes unreachable files in `.jj/repo/store/extra` to save
+  disk space.
+
 ### Fixed bugs
 
 * Broken symlink on Windows. [#6934](https://github.com/jj-vcs/jj/issues/6934).